New model of the backend is going on

+++ src/main/java/backend/project/repositories/Carrera_EstudianteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Carrera_EstudianteRepository extends JpaRepository<Carrera_Estudiante, Integer> {
+public interface Carrera_EstudianteRepository extends JpaRepository<Carrera_Estudiante, Long> {
 }
